Push del sprint 2 parcial (no documentado)

+++ InnovaSolutions/Vistas/Teacher/EditarExamen.aspx.cs
@@ -28,6 +28,16 @@ namespace InnovaSolutions.Vistas.Teacher
             test = Request.QueryString["Exam"];
             LoadCategories();
             LoadExam();
+
+            Image1.ImageUrl = "/Imagenes/Sony.png";
+            if (Conexion.Membresia_Conectado == "Gratis")
+            {
+                Image1.Visible = true;
+            }
+            else
+            {
+                Image1.Visible = false;
+            }
         }
 
         protected void btn_finish_Click(object sender, EventArgs e)
@@ -79,44 +89,54 @@ namespace InnovaSolutions.Vistas.Teacher
         {
             try
             {
-                string cant = txt_cant.Value;
-                if (string.IsNullOrWhiteSpace(cant)) { cant = "0"; }
-
-                SqlCommand com;
-                string ph = "Place Holder";
-
-                conex.Open();
-                com = conex.CreateCommand();
-                com.CommandText =
-                    "UPDATE Examen SET " +
-                    "Nombre = '" + txt_title.Text + "', " + 
-                    "Cantidad = " + cant + " " +
-                    "WHERE ID_Examen = " + test + "; " +
-                    "DELETE FROM Pregunta " +
-                    "WHERE ID_Examen = " + test;
-                com.ExecuteNonQuery();
-
-                foreach (string jsonGot in hf_hook.Value.Split('▄'))
+                if (hf_hook.Value.Equals(""))
                 {
-                    Pregunta p = JsonConvert.DeserializeObject<Pregunta>(jsonGot);
-                    string cmd =
-                        "INSERT INTO Pregunta VALUES (" +
-                        test + ", " +
-                        p.id_categoria + ", " +
-                        "'" + p.pregunta + "', " +
-                        "'" + string.Join(",", p.respuestas) + "', " +
-                        p.correcta + ", " +
-                        "\"" + p.grafico + "\");";
-
-                    System.Diagnostics.Debug.WriteLine(cmd);
-                    com.CommandText = cmd;
-                    com.ExecuteNonQuery();
+                    Response.Write("<script>alert('El valor regresado es nulo\n" +
+                        "introduzca los datos una vez mas');</script>\n" +
+                        "si esto sucede a menudo contactese con el proveedor");
                 }
+                else
+                {
+                    string cant = txt_cant.Value;
+                    if (string.IsNullOrWhiteSpace(cant)) { cant = "0"; }
 
-                com.Cancel();
-                conex.Close();
-                Response.Write("<script>alert('Done');</script>");
-                Response.Redirect("MenuExams.aspx");
+                    SqlCommand com;
+                    string ph = "Place Holder";
+
+                    conex.Open();
+                    com = conex.CreateCommand();
+                    com.CommandText =
+                        "UPDATE Examen SET " +
+                        "Nombre = '" + txt_title.Text + "', " +
+                        "Cantidad = " + cant + " " +
+                        "WHERE ID_Examen = " + test + "; " +
+                        "DELETE FROM Pregunta " +
+                        "WHERE ID_Examen = " + test;
+                    com.ExecuteNonQuery();
+
+                    foreach (string jsonGot in hf_hook.Value.Split('▄'))
+                    {
+                        Pregunta p = JsonConvert.DeserializeObject<Pregunta>(jsonGot);
+                        string cmd =
+                            "SET QUOTED_IDENTIFIER OFF " +
+                            "INSERT INTO Pregunta VALUES (" +
+                            test + ", " +
+                            p.id_categoria + ", " +
+                            "'" + p.pregunta + "', " +
+                            "'" + string.Join(",", p.respuestas) + "', " +
+                            p.correcta + ", " +
+                            "\"" + p.grafico + "\");";
+
+                        System.Diagnostics.Debug.WriteLine(cmd);
+                        com.CommandText = cmd;
+                        com.ExecuteNonQuery();
+                    }
+
+                    com.Cancel();
+                    conex.Close();
+                    Response.Write("<script>alert('Done');</script>");
+                    Response.Redirect("MenuExams.aspx");
+                }
             }
             catch (Exception e)
             {
